cleanup file: zboxcli_collaborator_test.go
